requirements.txt
SpeechRecognition
pocketsphinx
pyaudio
google-cloud-speech
google-cloud-texttospeech
pygame
sounddevice
playsound
